branch v1.0.0
- Added basic validation for postUser controller
