+++ database/database.go
@@ -1,5 +1,29 @@
 package database
 
-import "go.mongodb.org/mongo-driver/mongo"
+import (
+	"context"
+	"log"
+	"time"
+
+	"go.mongodb.org/mongo-driver/mongo"
+	"go.mongodb.org/mongo-driver/mongo/options"
+)
+
+func CreateMongoClient(connectionString string) *mongo.Client {
+	client, err := mongo.NewClient(options.Client().ApplyURI(connectionString))
+	if err != nil {
+		log.Fatal(err)
+	}
+	return client
+}
+
+func ConnectMongo(client *mongo.Client) {
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	err := client.Connect(ctx)
+	if err != nil {
+		log.Fatal(err)
+	}
+}
 
 var DBConn *mongo.Client
