+++ helpers/validator.go
@@ -1,0 +1,7 @@
+package helpers
+
+import (
+	"github.com/go-playground/validator/v10"
+)
+
+var Validate *validator.Validate
